+++ src/emulator/portUtils.ts
@@ -205,9 +205,12 @@ const EMULATOR_CAN_LISTEN_ON_PRIMARY_ONLY: Record<PortName, boolean> = {
   functions: true,
   logging: true,
   storage: true,
+  tasks: true,
 
   // Only one hostname possible in .server mode, can switch to middleware later.
   hosting: true,
+
+  apphosting: true,
 };
 
 export interface EmulatorListenConfig {