+++ src/firebaseConfig.ts
@@ -203,6 +203,10 @@ export type EmulatorsConfig = {
     host?: string;
     port?: number;
   };
+  apphosting?: {
+    host?: string;
+    port?: number;
+  };
   pubsub?: {
     host?: string;
     port?: number;
@@ -231,6 +235,10 @@ export type EmulatorsConfig = {
   };
   singleProjectMode?: boolean;
   dataconnect?: {
+    host?: string;
+    port?: number;
+  };
+  tasks?: {
     host?: string;
     port?: number;
   };