+++ firebase-vscode/CHANGELOG.md
@@ -1,5 +1,15 @@
 ## NEXT
 
+## 0.8.0
+
+- Updated internal firebase-tools dependency to 13.17.0
+
+- [Fixed] Extension properly picks up firebase.json changes during Firebase Init flow
+
+## 0.7.0
+
+- Updated internal firebase-tools dependency to 13.16.0
+
 ## 0.6.2
 
 - Updated internal firebase-tools dependency to 13.15.4